arm64 deploy tests (#8160)
* attempt #100000 to do arm64 builds

* more debugging info

* remove --insecure

* remove / after pachyderm

* push_docker: fix lint

* goreleaser: don't build darwin pachctl when making images

* integration-tests: wait for docker images to be pushed

* wait for docker images in the deploy/rootless workflows

* wait_for_docker_images: fix env var

* docker-build: tag amd64 versions of pachd as latest and local; these aren't pushed but are usefully locally

* try running a deployment on arm64

* skip kubeval for arm64

* name deploy jobs

* try with etcd and postgres arm64

* temp turn off upgrade tests

* try postgres image

* okay try this one

* wrong override perhaps

* disable loki

* try arm64 friendly values

* disable kube event tail for now

* disable loki option

* reenable enterprise test

* try a multi arch kube-even-tail image

* test with new image

* use build tags to skip non arm64 tests

* test 1.1.6.1 pgbouncer

* Revert "use build tags to skip non arm64 tests"

This reverts commit 4e55601100f144b3c4cc24efb68bdd79a44828c2.

* test pgbouncer 1.17.0

* Revert "test pgbouncer 1.17.0"

This reverts commit 8795a16931ee8fc8c15e77ce98c7ff0ca1014d6c.

* bump kube-event-tail to arm64 compatible image

* skip upgrade deploy tests for arm64

* overrides only for arm64 deploy tests

+++ etc/testing/circle/install.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-set -ex
+set -euxo pipefail
+
+ARCH=amd64
+if [ "$(uname -m)" = "aarch64" ]; then ARCH=arm64; fi
 
 mkdir -p cached-deps
 
@@ -26,7 +29,7 @@ sudo chown root:root /etc/fuse.conf
 # curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt
 if [ ! -f cached-deps/kubectl ] ; then
     KUBECTL_VERSION=v1.23.5
-    curl -L -o kubectl https://storage.googleapis.com/kubernetes-release/release/${KUBECTL_VERSION}/bin/linux/amd64/kubectl && \
+    curl -L -o kubectl https://storage.googleapis.com/kubernetes-release/release/${KUBECTL_VERSION}/bin/linux/${ARCH}/kubectl && \
         chmod +x ./kubectl
         mv ./kubectl cached-deps/kubectl
 fi
@@ -35,16 +38,16 @@ fi
 # To get the latest minikube version:
 # curl https://api.github.com/repos/kubernetes/minikube/releases | jq -r .[].tag_name | sort -V | tail -n1
 if [ ! -f cached-deps/minikube ] ; then
-    MINIKUBE_VERSION=v1.26.0 
-    curl -L -o minikube https://storage.googleapis.com/minikube/releases/${MINIKUBE_VERSION}/minikube-linux-amd64 && \
+    MINIKUBE_VERSION=v1.26.0
+    curl -L -o minikube https://storage.googleapis.com/minikube/releases/${MINIKUBE_VERSION}/minikube-linux-${ARCH} && \
         chmod +x ./minikube
         mv ./minikube cached-deps/minikube
 fi
 
 # Install kubeval
-if [ ! -f cached-deps/kubeval ]; then
+if [ ! -f cached-deps/kubeval ] && [ "$ARCH" = "amd64" ]; then
   KUBEVAL_VERSION=0.15.0
-  curl -L https://github.com/instrumenta/kubeval/releases/download/${KUBEVAL_VERSION}/kubeval-linux-amd64.tar.gz \
+  curl -L https://github.com/instrumenta/kubeval/releases/download/${KUBEVAL_VERSION}/kubeval-linux-${ARCH}.tar.gz \
       | tar xzf - kubeval
       mv ./kubeval cached-deps/kubeval
 fi
@@ -52,9 +55,9 @@ fi
 # Install helm
 if [ ! -f cached-deps/helm ]; then
   HELM_VERSION=3.5.4
-  curl -L https://get.helm.sh/helm-v${HELM_VERSION}-linux-amd64.tar.gz \
-      | tar xzf - linux-amd64/helm
-      mv ./linux-amd64/helm cached-deps/helm
+  curl -L https://get.helm.sh/helm-v${HELM_VERSION}-linux-${ARCH}.tar.gz \
+      | tar xzf - linux-${ARCH}/helm
+      mv ./linux-${ARCH}/helm cached-deps/helm
 fi
 
 # Install goreleaser
